+++ ui/mods/keybind_extensions/keybinds.js
@@ -68,6 +68,8 @@
 
   set = 'gameplay'
   display_sub_group = 'Selection Edit'
+  kb('only_construction_in_selection')
+  kb('remove_construction_from_selection')
   kb('only_bots_in_selection')
   kb('remove_bots_from_selection')
   kb('only_tanks_in_selection')